+++ lib/src/ui/pages/cupertino/list_section.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'list_section.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$configHash() => r'0ce0181e092b5d18ed61c25c51b33a05dd430565';
+
+/// See also [Config].
+@ProviderFor(Config)
+final configProvider =
+    AutoDisposeNotifierProvider<Config, ListSectionConfig>.internal(
+  Config.new,
+  name: r'configProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$configHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Config = AutoDisposeNotifier<ListSectionConfig>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member