+++ lib/src/ui/pages/cupertino/widgets.dart
@@ -1,5 +1,6 @@
 import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
 
+import 'list_section.dart';
 import 'list_tile.dart';
 import 'slider.dart';
 import 'switch.dart';
@@ -8,4 +9,5 @@ final cupertinoItems = <CodeItem>[
   switchItem,
   sliderItem,
   listTileItem,
+  listSectionItem,
 ];